+++ app/controllers/external_controller.rb
@@ -46,6 +46,10 @@ class ExternalController < ApplicationController
   # def thanks
   # end
 
+  # # GET /privacy-policy
+  # def privacy_policy
+  # end
+
   # # GET /version
   # # GET /version.json
   # def version