+++ config/initializers/themes.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+BostonSleep::THEMES = [
+  ["Default", "default"]
+]