+++ test/system/external_test.rb
@@ -63,4 +63,14 @@ class ExternalTest < ApplicationSystemTestCase
     click_form_submit
     screenshot("visit-contact-page")
   end
+
+  test "visit the privacy policy page" do
+    visit privacy_policy_url
+    screenshot("visit-privacy-policy-page")
+  end
+
+  test "visit the version page" do
+    visit version_url
+    screenshot("visit-version-page")
+  end
 end